telegram_bot/requirements.txt
peewee~=3.16.2
requests~=2.31.0
python-dotenv~=1.0.0
pydantic~=1.10.9
python-telegram-bot~=20.3